Support subprojects in "infoplist-modify" task
When the project using the xcode plugin is part of a multi-project Gradle build,
the infoplist-modify task fails with file not found error from 'plutil'.
The cause is that the generated path to the .pbxproj file is missing the subproject's directory.
For example, the path /home/username/project/foo.xcodeproj/project.pbxproj was generated
instead of the correct path /home/username/project/subproject/foo.xcodeproj/project.pbxproj

This change adds the subdirectory to the path.
Also tested successfully with a single-project build (no change in path there).