Support subprojects in "test" task
When the project using the xcode plugin is part of a multi-project Gradle build,
the test task fails with file not found error from 'xcodebuild'.
The cause is that the 'xcodebuild' command for tests is run in the current directory.
In a multi-project build the current directory will be the parent directory of the project
when executing a build from the parent project, and not the subproject's directory.

This change sets the current working directory for the test task's 'xcodebuild' command to be the current project's directory.
Also tested successfully with a single-project build (no change in path there).