Common task name prefix support
Common task names like build, test and archive are used in many Gradle plugins.
Task names have to be unique, therefore such plugins cannot be used together in the
same project.

To solve this problem a new xcodePluginPrefix property is added to the
plugin to set the prefix of the common task names. This property has to be set in the
project before applying the Xcode plugin.