Fixing broken test XcodeBuildTaskTest.run_command_xcodeversion.
It is accessing the filesystem and it will break if you dont have /Applications/Xcode.app in your machine.

"mdfind" will only return paths that exist in the filesystem, so Im just removing the check on our side. If for some reason we still want to check that it exists, then we will need to mock/inject things somehow.